birdclef-2026/requirements-colab.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66
torch>=2.1
soundfile>=0.12
